# packages/contracts/openapi/runtime-gateway.openapi.yaml
openapi: 3.1.0
info:
  title: AgentTwin runtime gateway API
  version: 1.0.0
  summary: Runtime containment of an agent's tool actions — policies, approvals and every decision.
  description: |
    The runtime gateway is an opt-in path through which an agent's tool
    actions pass at runtime (spec §30-31, ADR-0033). Each call is decided by
    the tool's active policies: allowed, allowed with limits, sent to a
    person for approval, or denied — before it reaches the tool.

    * **Tool endpoints** name the tools the gateway may call, their risk tier
      and timeout; a URL outside the egress allowlist is refused.
    * **Policies** are named, versioned documents of CEL rules that guard one
      tool. A version is tested (its own tests and every numeric threshold)
      and activated only when every test passes.
    * **Approvals** bind a person's decision to one exact action: the approved
      action runs once; a changed action needs its own approval.
    * **Decisions** record every call the gateway decided (append-only).

    **Where to call it.** The public API is served by the control plane (the
    edge), which authenticates the caller, applies rate limits and, on
    `/api/v1` mutations, `Idempotency-Key`; it forwards the request to the
    runtime gateway. The invocation path (`/gateway/v1`) is reached the same
    way with an API key holding the `runtime:invoke` scope; there the
    gateway itself implements `Idempotency-Key`, per action, and forwards the
...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKey: []
paths:
  /api/v1/tool-endpoints:
    get:
      operationId: listToolEndpoints
      tags: [Tool endpoints]
      summary: List tool endpoints
      description: The project's registered tools, by name.
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
      responses:
        '200':
          description: The registered tools.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolEndpointPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/tool-endpoints/{tool}:
    parameters:
      - $ref: '#/components/parameters/Tool'
    get:
      operationId: getToolEndpoint
      tags: [Tool endpoints]
      summary: Read a tool endpoint
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
      responses:
        '200':
          description: The tool endpoint.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolEndpoint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
    put:
      operationId: putToolEndpoint
      tags: [Tool endpoints]
      summary: Register or replace a tool endpoint
      description: |
        Registers the tool (`201`) or replaces its endpoint (`200`); needs
        `policy.write`. The URL's host must be in the egress allowlist
        (`RUNTIME_EGRESS_ALLOWED_HOSTS`, or `RUNTIME_EGRESS_PRIVATE_HOSTS` for
        an internal service); every call is checked again when it dials.
        An irreversible tool requires an idempotency key unless
        `idempotency: optional` is set. Audited (`tool_endpoint.registered` or
        `tool_endpoint.updated`).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolEndpointInput'
      responses:
        '200':
          description: The endpoint was replaced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolEndpoint'
        '201':
          description: The tool was registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolEndpoint'
        '400':
          $ref: '#/components/responses/EndpointInvalid'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyInProgress'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
    delete:
      operationId: deleteToolEndpoint
      tags: [Tool endpoints]
      summary: Remove a tool endpoint
      description: |
        The tool can no longer be called through the gateway (`404
        TOOL_NOT_REGISTERED`); its policies and decisions are kept. Needs
        `policy.write`; audited (`tool_endpoint.removed`).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Removed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyInProgress'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/policies:
    get:
      operationId: listPolicies
      tags: [Policies]
      summary: List policies
      description: The project's policies by name, optionally those guarding one tool.
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
        - name: tool
          in: query
          schema:
            $ref: '#/components/schemas/ToolName'
      responses:
        '200':
          description: The policies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
    post:
      operationId: createPolicy
      tags: [Policies]
      summary: Create a policy
      description: |
        Stores a policy document (YAML or JSON, `apiVersion:
        agenttwin.dev/v1`, `kind: Policy`) as version 1 of a new policy named
        by its `metadata.name`, guarding `spec.tool`. Every rule is compiled
        and type-checked; nothing is activated. Needs `policy.write`; audited
        (`policy.created`).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyDocumentInput'
      responses:
        '201':
          description: The policy and its first version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyCreated'
        '400':
          $ref: '#/components/responses/PolicyInvalid'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/PolicyExists'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/policies/test:
    post:
      operationId: testPolicy
      tags: [Policies]
      summary: Test a policy
      description: |
        Decides a draft document, or a saved version, against its own tests
        and the extra `cases`, and around every numeric threshold its rules
        compare against: `args.amount > 100` is decided at 99, 99.99, 100,
        100.01 and 101 (one hundredth either side of a decimal threshold; one
        call either side of `trace.calls`). The boundaries start from `base`,
        else from the first test's action. Reports the rules no test decided
        and whether the version could be activated for the tool as
        registered. Needs `policy.test`; nothing is stored.
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyTestInput'
      responses:
        '200':
          description: What the policy decided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyTestReport'
        '400':
          $ref: '#/components/responses/PolicyTestInvalid'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyInProgress'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/policies/{policy_id}:
    parameters:
      - $ref: '#/components/parameters/PolicyId'
    get:
      operationId: getPolicy
      tags: [Policies]
      summary: Read a policy
      description: The policy, its active version and every version (newest first).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
      responses:
        '200':
          description: The policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/policies/{policy_id}/versions:
    parameters:
      - $ref: '#/components/parameters/PolicyId'
    post:
      operationId: addPolicyVersion
      tags: [Policies]
      summary: Add a policy version
      description: |
        Stores the document as the policy's next version (`201`). Its
        `metadata.name` and `spec.tool` must stay the policy's. A document
        that decides exactly what the latest version decides (the same
        `spec_hash`) returns that version (`200`, `created: false`). Versions
        are immutable; the active one does not change. Needs `policy.write`;
        audited (`policy.version_created`).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyDocumentInput'
      responses:
        '200':
          description: The latest version already decides this.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionAdded'
        '201':
          description: The new version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionAdded'
        '400':
          $ref: '#/components/responses/PolicyInvalid'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyInProgress'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/policies/{policy_id}/versions/{version}:
    parameters:
      - $ref: '#/components/parameters/PolicyId'
      - name: version
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
    get:
      operationId: getPolicyVersion
      tags: [Policies]
      summary: Read a policy version
      description: |
        The document as written and as read (defaults filled), the variables
        its rules can use and the numeric thresholds they compare against.
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
      responses:
        '200':
          description: The version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/policies/{policy_id}/activate:
    parameters:
      - $ref: '#/components/parameters/PolicyId'
    post:
      operationId: activatePolicy
      tags: [Policies]
      summary: Activate a policy version
      description: |
        Makes the version the policy's active one, from the next call on.
        Refused (`422 POLICY_NOT_ACTIVATABLE`, `details.problems`) unless the
        version has tests and every test passes; `fail_open` only for a tool
        registered as `READ`. Activating the active version changes nothing.
        Needs `policy.activate`; publishes `policy.activated.v1` and is
        audited (`policy.activated`).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateInput'
      responses:
        '200':
          description: The policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyInProgress'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/NotActivatable'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/policies/{policy_id}/deactivate:
    parameters:
      - $ref: '#/components/parameters/PolicyId'
    post:
      operationId: deactivatePolicy
      tags: [Policies]
      summary: Deactivate a policy
      description: |
        The policy no longer guards its tool, from the next call on. A reason
        is required: removing a guard is audited (`policy.deactivated`).
        Needs `policy.activate`.
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonInput'
      responses:
        '200':
          description: The policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '400':
          $ref: '#/components/responses/ReasonInvalid'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyInProgress'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/approvals:
    get:
      operationId: listApprovals
      tags: [Approvals]
      summary: List approval requests
      description: |
        Newest first. `status` is the status as of now: `EXPIRED` includes
        open requests past their expiry, which `PENDING` and `APPROVED`
        exclude.
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ApprovalStatus'
        - name: tool
          in: query
          schema:
            $ref: '#/components/schemas/ToolName'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/approvals/{approval_id}:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    get:
      operationId: getApproval
      tags: [Approvals]
      summary: Read an approval request
      description: |
        What the agent is trying to do (the tool, its risk tier, the redacted
        arguments), why it needs approval (the policy, rule and reason), the
        decision that asked, the expiry, the trace, and every call that
        presented the approval's token with how its arguments differ from the
        approved ones (spec §41.9).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/Project'
      responses:
        '200':
          description: The request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/approvals/{approval_id}/approve:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    post:
      operationId: approveApproval
      tags: [Approvals]
      summary: Approve the action
      description: |
        Approves this exact action, with a reason. The agent that asked then
        claims a token and the action runs once. A request that is no longer
        pending answers `409 APPROVAL_CLOSED` (`details.status`); one past its
        expiry is closed as `EXPIRED`. Needs `approval.decide`; audited
        (`approval.approved`).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonInput'
      responses:
        '200':
          description: The request, approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        '400':
          $ref: '#/components/responses/ReasonInvalid'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ApprovalClosed'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/approvals/{approval_id}/deny:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    post:
      operationId: denyApproval
      tags: [Approvals]
      summary: Deny the action
      description: |
        Denies this action, with a reason; the agent's token claim answers
        `403 APPROVAL_DENIED`. The same rules as approving. Needs
        `approval.decide`; audited (`approval.denied`).
      parameters:
        - $ref: '#/components/parameters/Organization'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonInput'
      responses:
        '200':
          description: The request, denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        '400':
          $ref: '#/components/responses/ReasonInvalid'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ApprovalClosed'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/IdempotencyKeyReused'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/Internal'
        '502':
          $ref: '#/components/responses/UpstreamUnavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A session token (dev login), an OIDC access token or an API key.
    apiKey:
      type: apiKey
      in: header
      name: X-AgentTwin-Api-Key
      description: A project API key.
  parameters:
...
    Organization:
      name: X-AgentTwin-Org
      in: header
      required: false
      description: The organization to act in, for credentials that belong to several.
      schema:
        $ref: '#/components/schemas/Uuid'
    Project:
      name: project_id
      in: query
      required: true
      description: The project. A project outside the caller's access answers `404`.
      schema:
        $ref: '#/components/schemas/Uuid'
...
  responses:
...
    Unauthenticated:
      description: No valid credential (`UNAUTHENTICATED`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    ToolEndpointPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ToolEndpoint'
...